cloud/requirements.txt
click~=8.1.3
PyYAML~=6.0
schema~=0.7.5
Jinja2~=3.1.2
paramiko~=2.11.0
scp~=0.14.4
pycryptodome~=3.15.0
validators~=0.20.0
docker~=6.0.0
boto3~=1.24.60
GitPython~=3.1.27